moved StartSpan/EndSpan to driver/internal
