+++ go/adbc/driver/internal/shared_utils.go
@@ -27,6 +27,9 @@ import (
 	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/array"
 	"github.com/apache/arrow-go/v18/arrow/memory"
+	"go.opentelemetry.io/otel/codes"
+	semconv "go.opentelemetry.io/otel/semconv/v1.30.0"
+	"go.opentelemetry.io/otel/trace"
 )
 
 const (
@@ -740,4 +743,34 @@ func ToXdbcDataType(dt arrow.DataType) (xdbcType XdbcDataType) {
 	default:
 		return XdbcDataType_XDBC_UNKNOWN_TYPE
 	}
+}
+
+// Starts a trace.Span with the given spanName for the tracing object with
+// the given ctx context.
+func StartSpan(ctx context.Context, spanName string, tracing adbc.OTelTracing, opts ...trace.SpanStartOption) (context.Context, trace.Span) {
+	if tracing == nil {
+		return ctx, trace.SpanFromContext(ctx)
+	}
+
+	attrs := tracing.GetInitialSpanAttributes()
+	attrs = append(attrs, semconv.DBOperationName(spanName))
+	opts = append(opts, trace.WithAttributes(attrs...))
+
+	return tracing.StartSpan(ctx, spanName, opts...)
+}
+
+// Ends the given span. If err is not nil, then the
+// error is recorded and the status is set appropriately.
+// Otherwise, the status is set to Ok.
+func EndSpan(span trace.Span, err error, options ...trace.SpanEndOption) {
+	if err != nil {
+		span.RecordError(err)
+		if adbcError, ok := err.(adbc.Error); ok {
+			span.SetAttributes(semconv.ErrorTypeKey.String(adbcError.Code.String()))
+		}
+		span.SetStatus(codes.Error, err.Error())
+	} else {
+		span.SetStatus(codes.Ok, "")
+	}
+	span.End(options...)
 }
