+++ go/adbc/driver/snowflake/connection.go
@@ -35,7 +35,6 @@ import (
 	"github.com/apache/arrow-adbc/go/adbc"
 	"github.com/apache/arrow-adbc/go/adbc/driver/internal"
 	"github.com/apache/arrow-adbc/go/adbc/driver/internal/driverbase"
-	"github.com/apache/arrow-adbc/go/adbc/utils"
 	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/array"
 	"github.com/snowflakedb/gosnowflake"
@@ -171,8 +170,8 @@ func isWildcardStr(ident string) bool {
 }
 
 func (c *connectionImpl) GetObjects(ctx context.Context, depth adbc.ObjectDepth, catalog, dbSchema, tableName, columnName *string, tableType []string) (reader array.RecordReader, err error) {
-	ctx, span := utils.StartSpan(ctx, "GetObjects", c)
-	defer utils.EndSpan(span, err)
+	ctx, span := internal.StartSpan(ctx, "GetObjects", c)
+	defer internal.EndSpan(span, err)
 
 	var (
 		pkQueryID, fkQueryID, uniqueQueryID, terseDbQueryID string
@@ -621,8 +620,8 @@ func (c *connectionImpl) getStringQuery(query string) (value string, err error) 
 }
 
 func (c *connectionImpl) GetTableSchema(ctx context.Context, catalog *string, dbSchema *string, tableName string) (sc *arrow.Schema, err error) {
-	ctx, span := utils.StartSpan(ctx, "GetTableSchema", c)
-	defer utils.EndSpan(span, err)
+	ctx, span := internal.StartSpan(ctx, "GetTableSchema", c)
+	defer internal.EndSpan(span, err)
 
 	tblParts := make([]string, 0, 3)
 	if catalog != nil {
